+++ cupy/testing/hypothesis_testing.py
